# openapi.yaml
openapi: 3.0.3
info:
  title: Tokabu Town API
  description: API for the Tokabu Town game - MVP Town-Investment Loop
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/dungeon/character/{characterId}/unclaimed:
    get:
      tags: [Dungeon]
      summary: Get unclaimed dungeon runs
      description: Get all unclaimed (finished but not claimed) dungeon runs for a character
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: string
          description: Character ID
      responses:
        '200':
          description: Unclaimed runs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/DungeonRun'
                  count:
                    type: integer
                    description: Number of unclaimed runs
        '400':
          description: Missing character ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Character not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to get unclaimed dungeon runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/dungeon/claim:
    post:
      tags: [Dungeon]
      summary: Claim dungeon run reward
      description: Claim reward for a specific completed dungeon run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dungeonRunId, userId]
              properties:
                dungeonRunId:
                  type: string
                  description: ID of the dungeon run to claim
                userId:
                  type: string
                  description: ID of the user claiming the reward
            example:
              dungeonRunId: run-123
              userId: user-456
      responses:
        '200':
          description: Reward claimed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reward claimed successfully
                  tokensAwarded:
                    type: number
                    description: Number of tokens awarded
                  run:
                    $ref: '#/components/schemas/DungeonRun'
        '400':
          description: Invalid claim request (already claimed, not finished, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Dungeon run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to claim reward
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/dungeon/claim-all:
    post:
      tags: [Dungeon]
      summary: Claim all dungeon run rewards
      description: Claim rewards for all unclaimed dungeon runs for a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [characterId, userId]
              properties:
                characterId:
                  type: string
                  description: ID of the character whose rewards to claim
                userId:
                  type: string
                  description: ID of the user claiming the rewards
            example:
              characterId: char-123
              userId: user-456
      responses:
        '200':
          description: All rewards claimed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: All rewards claimed successfully
                  totalTokens:
                    type: number
                    description: Total number of tokens awarded
                  claimedCount:
                    type: integer
                    description: Number of runs claimed
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/DungeonRun'
        '400':
          description: Invalid claim request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Character not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to claim all rewards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/dungeon/runs/{dungeonRunId}:
    get:
      tags: [Dungeon]
      summary: Get dungeon run by ID
      description: Get a specific dungeon run by its ID
      parameters:
        - name: dungeonRunId
          in: path
          required: true
          schema:
            type: string
          description: Dungeon run ID
      responses:
        '200':
          description: Dungeon run retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  run:
                    $ref: '#/components/schemas/DungeonRun'
        '400':
          description: Missing dungeon run ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Dungeon run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to get dungeon run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/dungeon/character/{characterId}/runs:
    get:
      tags: [Dungeon]
      summary: Get all character dungeon runs
      description: Get all dungeon runs (active, completed, and claimed) for a character
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: string
          description: Character ID
      responses:
        '200':
          description: Dungeon runs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/DungeonRun'
                  count:
                    type: integer
                    description: Total number of runs
        '400':
          description: Missing character ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Character not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to get character dungeon runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    DungeonRun:
      type: object
      properties:
        id:
          type: string
          description: Dungeon run UUID
        character_id:
          type: string
          description: Character ID that ran the dungeon
        user_id:
          type: string
          description: User ID that owns the character
        duration_seconds:
          type: integer
          description: Duration of the dungeon run in seconds
        starting_damage_rating:
          type: integer
          description: Character's damage rating at the start of the run
        started_at:
          type: string
          format: date-time
          description: When the dungeon run started
        finished_at:
          type: string
          format: date-time
          nullable: true
          description: When the dungeon run finished (null if still active)
        claimed_at:
          type: string
          format: date-time
          nullable: true
          description: When the rewards were claimed (null if not yet claimed)
        reward_amount:
          type: number
          nullable: true
          description: Amount of tokens rewarded (null if not yet calculated)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time